Struct google_safebrowsing4::ThreatEntry

pub struct ThreatEntry {
    pub url: Option<String>,
    pub hash: Option<String>,
    pub digest: Option<String>,
}

An individual threat; for example, a malicious URL or its hash representation. Only one of these fields should be set.

This type is not used in any activity, and only used as part of another schema.

Fields
url: Option<String>

A URL.

hash: Option<String>

A hash prefix, consisting of the most significant 4-32 bytes of a SHA256 hash. This field is in binary format.

digest: Option<String>

The digest of an executable in SHA256 format. The API supports both binary and hex digests.