Struct google_safebrowsing4::ThreatMatch

pub struct ThreatMatch {
    pub threat_type: Option<String>,
    pub threat: Option<ThreatEntry>,
    pub threat_entry_metadata: Option<ThreatEntryMetadata>,
    pub threat_entry_type: Option<String>,
    pub platform_type: Option<String>,
    pub cache_duration: Option<String>,
}

A match when checking a threat entry in the Safe Browsing threat lists.

This type is not used in any activity, and only used as part of another schema.

Fields
threat_type: Option<String>

The threat type matching this threat.

threat: Option<ThreatEntry>

The threat matching this threat.

threat_entry_metadata: Option<ThreatEntryMetadata>

Optional metadata associated with this threat.

threat_entry_type: Option<String>

The threat entry type matching this threat.

platform_type: Option<String>

The platform type matching this threat.

cache_duration: Option<String>

The cache lifetime for the returned match. Clients must not cache this response for more than this duration to avoid false positives.